Optimize resource definitions in template
Replaced the separate policy, and role definitions because it wasn't an
efficient way of implementing them, especially in consideration of the
lambda that needs them. With this commit, the role is tied to the
function, and the policy is associated to the role directlry without
having to use a reference.

Using the paramters from the build, naming the lambda, role, and policy
is easy, in addition to adding a path to conform to InfoSec standards.