Add ability to adjust CFN through Makefile
The initial revision of the Makefile did not support extensive
customization of the account target, nor the public bucket options.
This commit adds the functionality by supporting the definition of
specific variables, easing the deployment of custom S3 public bucket
policies across a number of accounts without having to make changes to
the CloudFormation template.