Replace cfn-lambda-handler with crhelper
While researching on ways to improve this project, I discovered a
custom resource helper developed by Amazonians, licensed with Apache v2
instead of the ISC license in use by the other library. In addition,
the new library handles the status code returned to the CloudFormation
endpoint without having to write the logic in the handler code.