Add policy for Lambda Execution Role, and KMS
Adding the AWSLambdaBasicExecutionRole ensures that the function will
have the permissions to create a log group, and put events into the
target log group.

To support the use of encrypted properties within Lambda, the execution
role requires access to the default key in KMS with the alias
aws/lambda to support decryption of the properties.