Add dependency for the role to the application
I noticed a significant lag from when the role is created, to when it's
available to the Lambda service for usage. This variable ensures that
CloudFormation waits for the Role to be created before creating the
Serverless Application.

Despite this, I added a sleep in the Makefile before triggering the use
of the application through CloudFormation.