Refactor solution to enable policy update via CFN
Before this commit, it was not possible to update the S3 control policy
simply by passing in different values for the four different polices in
the make build. This commit moves the policy values from the template
CFN to the stack as variables. These variables are passed in the event
from CloudFormation to Lambda (super awesome!), which are then consumed
by the lambda in the handler event.